ver 1.1
The first okay version...
Add more items to the GUI, though the function is still limited.
There are still a few conditions which are not dealt with well in this version.

The code is kind of redundant, especially in the plot_gui. Actually I should write a class to make it more concise.